Added support for named destinations. Added support for unicode.
git-svn-id: http://pdfviewer-win32.googlecode.com/svn/trunk@5 1400ccdc-6c00-11de-b131-07d8e4d3762e

+++ PDFLibNet/OutlineItem.h
@@ -35,7 +35,7 @@ namespace PDFLibNet
 		property LinkDest ^Destination{
 			LinkDest ^get(){
 				if(_dest==nullptr)
-					_dest =gcnew LinkDest((int)_outline->getDest());
+					_dest =gcnew LinkDest(_outline->getDest());
 				
 				return _dest;
 			}
